+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `laravel-multi-database-commands` will be documented in this file.
 
+## 1.1.0 - 2021-04-22
+
+- Refactory code to be able to test the package
+- Display the php artisan command that is been executed
+
 ## 1.0.0 - 2021-04-21
 
 - Execute any migration command (status, rollback, etc) on single or multi DB